python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "navidrome-python-stats-frontend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic-settings",
]

[tool.hatch.build.targets.wheel]
packages = ["navidrome_client", "web_server"]

[tool.uvicorn]
app = "main:app"
reload = true

